requirements.txt
asgiref==3.6.0
atomicwrites==1.4.1
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.0.12
colorama==0.4.6
Django==3.2
django-filter==22.1
djangorestframework==3.12.4
idna==3.4
iniconfig==2.0.0
packaging==23.0
pluggy==0.13.1
py==1.11.0
pycodestyle==2.10.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2022.7.1
requests==2.26.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
urllib3==1.26.14
